requirements.txt
aiohttp>=2.3.10
asn1crypto>=0.24.0
async-timeout>=3.0.1
beautifulsoup4>=4.9.0
certifi>=2018.8.24
chardet>=3.0.4
cryptography>=2.6.1
discord>=1.0.1
discord.py>=1.3.3
entrypoints>=0.3
enum34>=1.1.10
future>=0.18.2
idna>=2.6
idna-ssl>=1.1.0
iso8601>=0.1.12
joblib>=0.14.1
keyring>=17.1.1
keyrings.alt>=3.1.1
lxml>=4.5.0
multidict>=4.7.5
numpy>=1.18.2
olefile>=0.46
opencv-python>=3.4.6.27
Pillow>=5.4.1
playsound>=1.2.2
pycrypto>=2.6.1
PyGObject>=3.30.4
PyMarkovChain>=1.8
python-apt>=1.8.4.1
requests>=2.21.0
scikit-learn>=0.22.2.post1
scipy>=1.4.1
SecretStorage>=2.3.1
serial>=0.0.97
six>=1.12.0
sklearn>=0.0
soupsieve>=2.0
ssh-import-id>=5.7
urllib3>=1.24.1
websockets>=8.1
yarl>=1.4.2
